Refactored button creation into separate file. 37 / 66 tests work.
